Added an game over screen
Game over activity pops up when timer runs out. The activity reports the score, and gives option to play again or return to menu.

Note: One time I pressed "reset" the instant the timer ran out, and it popped up the game over screen, but when I replayed, the game over screen popped up again when the count timer had not yet hit zero. I replayed again, and the same thing happened. I was not able to reproduce this after ending and starting the app again, but just keep in the back of our heads that this may happen again.