Implemented game loop (#3)
* Implemented game loop

Set up the game loop. With every loop, we get a random instruction from a list of instructions and then add the corresponding UI elements. If the proper input is received, we clear the UI elements, wait one second, and then loop again. Also made it so the time available to perform the actions scales with score, but we might want to adjust this scale in the future.

* Implemented changes Fred requested

Added enums for Instructions, GameMode, and Difficulty. Instruction list is now constructed based on GameMode. Timer now pauses when app is paused and resumed when it resumes. Also cleaned up a few other things suggested by Fred.

* Fixed a bug where "Play Again" caused crashing

* Swapped MainActivity to be the main activity