Update High Score Storage
So I took your suggestions and I like it a lot more this way. However, when I try to set the text in GameActivity, the app still keeps crashing...

Not sure what I'm doing wrong here, I tried placing it in a few areas and confirmed the text box id, but I can't seem to figure out why it keeps crashing. Any suggestions? Ill continue to play around with it